requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
peewee==3.17.9
psycopg2-binary==2.9.10
pycryptodome==3.22.0
pycryptodomex==3.22.0
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
